improved client build
