+++ demo/game/index.js
@@ -4,9 +4,9 @@ const spawnConfig = { stdio: "inherit" };
 
 spawn(cmd, ["run", "game:server"], spawnConfig);
 
-setTimeout(() => {
-  spawn(cmd, ["run", "game:client"], spawnConfig);
-  spawn(cmd, ["run", "game:client"], spawnConfig);
-  spawn(cmd, ["run", "game:client"], spawnConfig);
-  spawn(cmd, ["run", "game:client"], spawnConfig);
-}, 250);
+// setTimeout(() => {
+//   spawn(cmd, ["run", "game:client"], spawnConfig);
+//   spawn(cmd, ["run", "game:client"], spawnConfig);
+//   spawn(cmd, ["run", "game:client"], spawnConfig);
+//   spawn(cmd, ["run", "game:client"], spawnConfig);
+// }, 250);
